Merge latest changes from Refactor branch

+++ src/fs.c
@@ -1,5 +1,24 @@
 #include "engine.h"
 
+#ifdef _WIN32
+    #include <io.h>
+    #define stat _stat64
+#else
+    #if !defined(_FILE_OFFSET_BITS)
+        #define _FILE_OFFSET_BITS 64
+    #endif
+#endif
+
+uint64_t get_file_size(const char *path)
+{
+    struct stat st;
+
+    if (stat(path, &st) != 0)
+        return (0);
+
+    return (uint64_t)st.st_size;
+}
+
 char *get_dirname(const char *path)
 {
     const char *last_slash = NULL;
@@ -90,6 +109,22 @@ cnbool is_dir(const char *path)
         struct stat s;
 
         return ((stat(path, &s) == 0 && S_ISDIR(s.st_mode)) ? true : false);
+    #endif
+}
+
+cnbool is_file(const char *path)
+{
+    #ifdef _WIN32
+        DWORD attr = GetFileAttributesA(path);
+
+        if (attr == INVALID_FILE_ATTRIBUTES)
+            return (false);
+
+        return ((attr & FILE_ATTRIBUTE_DIRECTORY) ? false : true);
+    #else
+        struct stat s;
+
+        return ((stat(path, &s) == 0 && S_ISREG(s.st_mode)) ? true : false);
     #endif
 }
 
@@ -340,7 +375,7 @@ uint8_t copy_file(const char *src, const char *dst)
     return (0);
 }
 
-uint8_t copytree(const char *src, const char *dst)
+uint8_t copytree(const char *src, const char *dst, cnbool overwrite)
 {
     #ifdef _WIN32
         WIN32_FIND_DATAA fd;
@@ -370,13 +405,20 @@ uint8_t copytree(const char *src, const char *dst)
             dst_path = join_path(dst, fd.cFileName);
 
             if (fd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
-                if (copytree(src_path, dst_path) != 0) {
+                if (copytree(src_path, dst_path, overwrite) != 0) {
                     FindClose(h);
                     free(src_path);
                     free(dst_path);
                     return (1);
                 }
             } else {
+                if (is_file(dst_path) && !overwrite) {
+                    FindClose(h);
+                    free(src_path);
+                    free(dst_path);
+                    continue;
+                }
+
                 if (copy_file(src_path, dst_path) != 0) {
                     FindClose(h);
                     free(src_path);
@@ -404,7 +446,7 @@ uint8_t copytree(const char *src, const char *dst)
         }
 
         dir = opendir(src);
-    
+
         if (!dir)
             return (1);
 
@@ -416,13 +458,18 @@ uint8_t copytree(const char *src, const char *dst)
             dst_path = join_path(dst, entry->d_name);
 
             if (is_dir(src_path)) {
-                if (copytree(src_path, dst_path) != 0) {
+                if (copytree(src_path, dst_path, overwrite) != 0) {
                     closedir(dir);
                     free(src_path);
                     free(dst_path);
                     return (1);
                 }
             } else {
+                if (is_file(dst_path) && !overwrite) {
+                    free(src_path);
+                    free(dst_path);
+                    continue;
+                }
                 if (copy_file(src_path, dst_path) != 0) {
                     closedir(dir);
                     free(src_path);
